Use less memory for small (<256) sets
Store indices in a byte[] when possible, switching between implementations on entry to the set code rather than in individual index-accessing methods to avoid the severe performance penalties we were seeing in the old compression code. Memory consumed creating a 6-element set is now 136B, versus 120B for the original compressed code (which was using nibbles for this size), 232B for the byte[]-using implementation this replaces, and 336B for a HashSet. A 49-element set (right before a resize) is at 1.1KB, versus 1.0KB/2.3KB/2.7KB respectively.