Discard timing outliers for faster, more accurate benchmarks
Instead of attempting to use JIT stats to restart measurement runs (which never seems to stabilise anyway), and then only sampling for a couple of seconds, discard outliers (more than three standard deviations above the mean), and continue to sample until the 99% confidence interval of the estimated mean is within 1% of it.

Outliers can be discarded retroactively if more than 3 standard deviations above the mean of the next 20 samples, as the worst outliers tend to come at the start of the run, when there are no statistics to go on. Outliers are discarded continuously to ensure the sample error is not biased upwards by them.

Mean and standard deviation are calculated off of exponentially weighted moving averages to increase the ability to spot larger initial outliers without setting the threshold too low and discarding accurate samples. The mean and s.d. used for sample error estimation and display still weights all (non-outlier) samples equally, however.

Benchmarks now typically take around 5s rather than 12s, and include many more samples.